my_streaming_project/consumer_service/requirements.txt
kafka-python
pyodbc